refactor(codegen): extract std::bytes::* into stdlib/bytes.rs (Round 1.17)
19 fns lifted into BytesStdlib<'ctx>: at_fallible, builder_new,
builder_append, builder_append_slice, builder_len, builder_finish,
builder_shift_front, builder_clear, builder_snapshot, builder_free,
builder_view, builder_text_view, is_alloc_fail, clone, from_string,
at, slice, from_int, concat + the lower_bytes_builder_handle_arg
helper used by every builder fn.

Surface bumps to pub(crate):
- Cx methods: emit_index_error_alloc, alloca_for
- FallibleCallResult struct fields (i1_path, out_val_slot, out_err_slot,
  success_ty, payload_ty) — bytes constructs the struct directly for
  its empty-bytes-on-error shape

codegen.rs: 38728 -> 37544 (-1184 lines).
